refactor: collateralization ratio row
